==> analysis/00_QC_checks.Rmd <==
---
title: "QC checks"
author: 
  - name: "Teun van den Brand"
    email: "t.vd.brand@nki.nl"
    affiliation: "Netherlands Cancer Institute"
date: '`r format(Sys.time(), "%d/%m/%Y")`'
output: 
  html_document:
    theme: united
    highlight: pygments
    fig_caption: yes
    code_folding: hide
    df_print: kable
    toc: true
    toc_depth: 4
    toc_float:
      collapsed: false
bibliography: /DATA/users/t.vd.brand/bibliography/articles.bibtex
knit: |
  # Redirect output to HTML folder
  (function(inputFile, encoding) {
    outputFile <- gsub("Rmd$", "html", basename(inputFile))
    rmarkdown::render(
      input = inputFile,
      encoding = encoding,
      output_file = here::here("html", outputFile)
    )
  })
---

## Setup {.tabset}

<!-- little HTML script to do indentation of the table of contents -->
<script>
    $(document).ready(function() {
      $items = $('div#TOC li');
      $items.each(function(idx) {
        num_ul = $(this).parentsUntil('#TOC').length;
        $(this).css({'text-indent': num_ul * 10, 'padding-left': 0});
      });

    });
</script>

```{css, echo = FALSE}
div.sourceCode {
  overflow-x: hidden;
}
```

### Load packages

```{r stfu, include = FALSE}
# No, I do *not* want your start-up messages, thank you very much
library <- function(...) {
  suppressPackageStartupMessages(base::library(...))
}
```


```{r load_packages}
library(here)
library(data.table)
library(GENOVA)
library(ggplot2)
library(scales)
```

### Load data

```{r load_data}
files <- list.files(
  here("data", "hic", "pools"),
  full.names = TRUE
)
names(files) <- tstrsplit(basename(files), "\\.")[[1]]

exps <- Map(
  load_contacts,
  signal_path = files,
  sample_name = names(files),
  MoreArgs = list(
    resolution = 5000000,
    scale_bp   = NULL,
    balancing  = FALSE
  )
)
exps <- sync_indices(exps)

config <- yaml::read_yaml(here("data", "hic", "project.yml"))

stat_files <- list.files(
  here("data", "hic", "stats_library"),
  full.names = TRUE
)
stats <- lapply(stat_files, function(file) {
  imp <- readLines(file, n = 8)
  imp <- tstrsplit(imp, "\t")
  imp <- setNames(as.numeric(imp[[2]]), imp[[1]])
  imp <- vctrs::data_frame(!!!imp)
  imp$file <- basename(file)
  imp
})
stats <- do.call(rbind, stats)
stats$sample <- tstrsplit(stats$file, "\\.")[[1]]
```

### Setup Aesthetics

```{r setup_aes}
mycolour <- "#000000FF" # Opaque Black

theme_set(theme_gray())
theme_update(text = element_text(colour = mycolour),
             line = element_line(colour = mycolour),
             aspect.ratio = 1,
             axis.line  = element_line(colour = mycolour),
             axis.ticks = element_line(colour = mycolour),
             axis.text  = element_text(colour = mycolour),
             legend.key = element_blank(),
             legend.background = element_rect(colour = NA, fill = NA),
             panel.background = element_blank(),
             panel.grid.major = element_line(colour = "grey95"),
             panel.grid.minor = element_blank(),
             plot.background  = element_blank(),
             strip.background = element_blank(),
             strip.text = element_text(colour = mycolour))
CYRUP <- c("#fed500", "#009bef", "#ff5c49", "#949494", "#D8D8D8")
rm(mycolour)
```

## Aim

## Library size

```{r}
libsize <- lapply(exps, function(exp) sum(exp$MAT$V3))

df <- data.frame(
  sample = names(libsize),
  size   = unlist(libsize),
  type   = tstrsplit(names(libsize), "_")[[2]]
)

ggplot(df, aes(size, sample)) +
  geom_col(aes(fill = type), colour = "black", 
           width = 2/(1+sqrt(5)), linewidth = 0.25) +
  geom_text(
    aes(size + 1e6, label = 
          number(size, scale = 1e-6, suffix = "M", accuracy = 0.1)),
    hjust = 0, size = 8.8 / .pt
  ) +
  coord_cartesian(clip = "off") +
  scale_x_continuous(
    labels = number_format(scale = 1e-6, suffix = "M"),
    expand = c(0, 0, 0.1, 0)
  )
```

## Cis-trans ratios

```{r}
cistrans <- cis_trans(exps)
cistrans$type <- tstrsplit(cistrans$sample, "_")[[2]]


ggplot(cistrans, aes(cis, sample)) +
  geom_col(aes(fill = type), colour = "black", 
           width = 2/(1+sqrt(5)), linewidth = 0.25) +
  geom_text(
    aes(cis + 1, label = number(cis, suffix = "%", accuracy = 0.1)),
    hjust = 0, size = 8.8 / .pt
  ) +
  coord_cartesian(clip = "off") +
  scale_x_continuous(
    expand = c(0, 0, 0.1, 0)
  )
```

```{r}
chr_mtx <- chromosome_matrix(exps, expected = "sums")
sum <- vapply(seq_len(dim(chr_mtx$obs)[3]), function(i) {
  x <- chr_mtx$obs[,,i]
  sum(x[col(x) != row(x)])
}, numeric(1))
obs <- vapply(seq_len(dim(chr_mtx$obs)[3]), function(i) {
  x <- chr_mtx$obs[,,i]
  diag(x) <- 0
  x / sum(x)
}, matrix(1, nrow = 23, ncol = 23))
# exp <- sweep(chr_mtx$exp, 3, sum, FUN = "*")
exp <- vapply(seq_len(dim(chr_mtx$exp)[3]), function(i) {
  x <- chr_mtx$exp[,,i]
  diag(x) <- 0
  x / sum(x)
}, matrix(1, nrow = 23, ncol = 23))

obsexp <- (obs - exp)

df <- reshape2::melt(obsexp)
df <- df[df$Var1 != df$Var2,]
df$Var3 <- factor(expnames(exps)[df$Var3], expnames(exps))

ggplot(df, aes(Var1, Var2, fill = value)) +
  geom_raster() +
  facet_wrap(~ Var3, ncol = 6) +
  scale_fill_GENOVA_div(midpoint = 0) +
  guides(
    x = guide_axis(angle=  90, check.overlap = TRUE, title = NULL),
    y = guide_axis(check.overlap = TRUE, title = NULL)
  )
```

```{r}
# Wrangle stats
swap <- stats$sample %in% c("P06_PE_rep2", "P08_PE")
stats$sample[swap] <- rev(stats$sample[swap])

repped <- stats[grepl("rep", stats$sample),]
repped$pool <- substr(repped$sample, 1, 6)
repped <- as.data.table(repped)
repped <- repped[, lapply(.SD, function(x) {
  if (is.numeric(x)) {
    sum(x)
  } else {
    x[1]
  }
}), by = "pool"]
repped$sample <- repped$pool
repped$pool <- NULL

new <- as.data.table(copy(stats))
new <- new[!grepl("rep", sample)]
new <- new[!(sample %in% repped$sample)]
new <- rbind(new, repped)

new[, proc_dups := total_dups / (total_dups + total_nodups)]

df <- tidyr::pivot_longer(new, c(total_dups, total_nodups))
df$label <- percent(df$proc_dups, accuracy = 0.1)
df$label[df$name == "total_nodups"] <- ""

ggplot(df, aes(value, sample, fill = name)) +
  geom_col(width = 2/(1 + sqrt(5))) +
  geom_text(aes(label = label),
            position = position_stack(vjust = 0.5),
            size = 3) +
  scale_x_continuous(
    expand = c(0, 0, 0.1, 0),
    labels = number_format(scale = 1e-6, suffix = "M"),
    name   = "Reads"
  ) +
  scale_fill_manual(
    values = CYRUP[3:2],
    name = "Type",
    labels = c("Duplicates", "Originals")
  ) +
  theme(
    panel.grid.major.y = element_blank(),
    axis.text.y.left = element_text(hjust = 0)
  )
```

## References

<div id="refs"></div>

## Session Info

```{r session_info, include=TRUE, echo=FALSE, results = "asis"}
pkgs <- loadedNamespaces()
libTvdB::format_sessioninfo_html(pkgs)
```
