==> analysis/03_01_TADs.Rmd <==
---
title: "TADs"
author: 
  - name: "Teun van den Brand"
    email: "t.vd.brand@nki.nl"
    affiliation: "Netherlands Cancer Institute"
date: '`r format(Sys.time(), "%d/%m/%Y")`'
output: 
  html_document:
    theme: united
    highlight: pygments
    fig_caption: yes
    code_folding: hide
    df_print: kable
    toc: true
    toc_depth: 4
    toc_float:
      collapsed: false
bibliography: /DATA/users/t.vd.brand/bibliography/articles.bibtex
knit: |
  # Redirect output to HTML folder
  (function(inputFile, encoding) {
    outputFile <- gsub("Rmd$", "html", basename(inputFile))
    rmarkdown::render(
      input = inputFile,
      encoding = encoding,
      output_file = here::here("html", outputFile)
    )
  })
---

## Setup {.tabset}

<!-- little HTML script to do indentation of the table of contents -->
<script>
    $(document).ready(function() {
      $items = $('div#TOC li');
      $items.each(function(idx) {
        num_ul = $(this).parentsUntil('#TOC').length;
        $(this).css({'text-indent': num_ul * 10, 'padding-left': 0});
      });

    });
</script>

```{css, echo = FALSE}
div.sourceCode {
  overflow-x: hidden;
}
```

### Load packages

```{r stfu, include = FALSE}
# No, I do *not* want your start-up messages, thank you very much
library <- function(...) {
  suppressPackageStartupMessages(base::library(...))
}
```


```{r load_packages}
library(here)
library(ggplot2)
library(scales)
library(GENOVA)
library(data.table)
library(GenomicRanges)
```

### Load data

```{r load_data}
files <- list.files(
  here("data", "hic", "pools"),
  full.names = TRUE
)
names(files) <- tstrsplit(basename(files), "\\.")[[1]]

centromeres <- readRDS(here("references", "centromeres.rds"))
```

### Setup Aesthetics

```{r setup_aes}
mycolour <- "#000000FF" # Opaque Black

theme_set(theme_gray())
theme_update(text = element_text(colour = mycolour),
             line = element_line(colour = mycolour),
             aspect.ratio = 1,
             axis.line  = element_line(colour = mycolour),
             axis.ticks = element_line(colour = mycolour),
             axis.text  = element_text(colour = mycolour),
             legend.key = element_blank(),
             legend.background = element_rect(colour = NA, fill = NA),
             panel.background = element_blank(),
             panel.grid.major = element_line(colour = "grey95"),
             panel.grid.minor = element_blank(),
             plot.background  = element_blank(),
             strip.background = element_blank(),
             strip.text = element_text(colour = mycolour))
CYRUP <- c("#fed500", "#009bef", "#ff5c49", "#949494", "#D8D8D8")
rm(mycolour)
```

## Aim

## Processing

```{r, message=FALSE}
results <- Map(function(file, nm) {
  
  # Load experiment
  exp  <- load_contacts(file, resolution = 20e3, sample_name = nm)
  
  # Calculate insulation score
  ins  <- insulation_score(exp)
  
  # Call TADs
  tads <- call_TAD_insulation(ins, min_strength = -Inf)
  
  # Filter tads
  filtered <- as.data.table(tads[[1]])
  filtered <- filtered[(end - start) > 225e3]
  filtered <- filtered[!overlapsAny(GPos(chrom, start), centromeres)]
  filtered <- filtered[!overlapsAny(GPos(chrom, end), centromeres)]
  
  # Make ATA
  zexp <- load_contacts(file, resolution = 20e3, z_norm = TRUE,
                        sample_name = nm)
  ata <- ATA(zexp, as.data.frame(filtered))
  
  # Give output
  list(
    insulation = ins,
    tads       = tads,
    ata        = ata
  )
}, file = files, nm = names(files))
```

### Extracting

```{r}
atas <- lapply(results, `[[`, "ata")
atas <- do.call(bundle, atas)
expnames(atas) <- names(files)

insulation <- lapply(results, function(res) res$insulation$insula_score)
insulation <- lapply(insulation, function(df) {
  colnames(df)[5] <- "value"
  df
})
insulation <- rbindlist(insulation, idcol = "exp")

tads <- lapply(lapply(results, `[[`, "tads"), `[[`, 1L)
tads <- rbindlist(tads, idcol = "exp")
```

## Insulation

```{r}
ins <- dcast(insulation, chrom + start + end ~ exp, value.var = "value")

scores <- as.matrix(as.data.frame(ins)[, -c(1:4)])
scores <- scores[rowSums(!is.finite(scores)) == 0, ]

pca <- prcomp(scale(t(scores)))
varexp <- pca$sdev^2 / sum(pca$sdev^2)

df <- data.frame(
  pc1 = pca$x[, 1],
  pc2 = pca$x[, 2],
  nm  = colnames(scores),
  pat  = tstrsplit(colnames(scores), "_")[[1]],
  type = tstrsplit(colnames(scores), "_")[[2]]
)

ggplot(df, aes(pc1, pc2, colour = type, label = pat)) +
  geom_text() +
  ggh4x::coord_axes_inside(ratio = 1) +
  # coord_equal() +
  labs(
    x = paste0("PC1 ", percent(varexp[1])),
    y = paste0("PC2 ", percent(varexp[2]))
  ) +
  theme(
    aspect.ratio = NULL
  )
```

## TADs

### Counts

```{r}
df <- tads
df$pat  <- tstrsplit(tads$exp, "_")[[1]]
df$type <- tstrsplit(tads$exp, "_")[[2]]

ggplot(df, aes(y = exp, fill = type)) +
  geom_bar()
```

### Sizes

```{r}
ggplot(df, aes(end - start, exp)) +
  geom_bin_2d() +
  scale_fill_viridis_c() +
  scale_x_continuous(
    trans = "log10",
    breaks = 10^c(4:7),
    labels = c("10kb", "100kb", "1Mb", "10Mb")
  )
```

### Border sharing

```{r}
borders <- GRanges(
  seqnames = c(tads$chrom, tads$chrom),
  IRanges(c(tads$start, tads$end), width = 1),
  exp = c(tads$exp, tads$exp)
)
wide_borders <- resize(borders, 3 * 20e3, fix = "center")
wide_borders <- reduce(wide_borders)

olap <- findOverlaps(borders, wide_borders)

df <- data.table(
  border_id = to(olap),
  exp = borders$exp[from(olap)]
)
df <- df[, .(n_sample = length(unique(exp))), by = "border_id"]
df <- as.data.frame(table(df$n_sample))

ggplot(df, aes(Var1, Freq)) +
  geom_col() +
  labs(
    x = "# Samples",
    y = "# Borders"
  ) +
  scale_y_continuous(
    expand = c(0, 0, 0.1, 0)
  )
```

#### Jaccard index

```{r}
binary <- vapply(
  split(borders, borders$exp),
  function(gr) {
    overlapsAny(wide_borders, gr)
  },
  logical(length(wide_borders))
)

cs <- colSums(binary)
int <- crossprod(binary)
uni <- outer(cs, cs, "+") - int

jac <- int / uni
jac <- reshape2::melt(jac)
jac <- subset(jac, Var1 != Var2)

ggplot(jac, aes(Var1, Var2, fill = value)) +
  geom_raster() +
  scale_fill_viridis_c() +
  guides(x = guide_axis(angle = 90))
```

##### Cluster

```{r}
dist  <- as.dist(1 - (int / uni))
clust <- hclust(as.dist(1 - (int / uni)))
clust <- seriation:::reorder.hclust(clust, dist)

ggplot(jac, aes(Var1, Var2, fill = value)) +
  geom_raster() +
  scale_fill_viridis_c() +
  scale_x_discrete(limits = clust$labels[clust$order]) +
  scale_y_discrete(limits = clust$labels[clust$order]) +
  guides(x = guide_axis(angle = 90))
```



## ATA

```{r}
df <- reshape2::melt(atas$signal)

ggplot(df, aes(Var1, Var2, fill = value)) +
  geom_raster() +
  facet_wrap(~ Var3, ncol = 6) +
  scale_fill_GENOVA_div(midpoint = 0, name = "Avg.\nZ-score") +
  scale_x_continuous(breaks = c(5, 15) * 1e5, 
                     labels = c("5'", "3'"),
                     name = NULL) +
  scale_y_continuous(breaks = c(5, 15) * 1e5, 
                     labels = c("3'", "5'"),
                     name = NULL)

```

### Quantification

```{r}
qata <- quantify(atas, metric = "mean")

ggplot(qata$per_TAD, aes(difference)) +
  geom_density(fill = "black") +
  scale_x_continuous(limits = c(-1, 2), oob = oob_keep) +
  scale_y_continuous(limits = c(0, 2), oob = oob_squish,
                     breaks = NULL) +
  facet_grid(sample ~ .) +
  
  theme(aspect.ratio = NULL,
        panel.spacing.y = unit(0, "pt"),
        strip.text.y = element_text(angle = 0, hjust = 0))
```

```{r}
atas$signal_raw <- NULL

saveRDS(as.data.frame(tads), here("rds", "01_TADs.rds"))
saveRDS(qata, here("rds", "01_ZATA_quantification.rds"))
saveRDS(atas, here("rds", "01_ZATA_signal.rds"))
saveRDS(insulation, here("rds", "01_insulation_score.rds"))
```


## References

<div id="refs"></div>

## Session Info

```{r session_info, include=TRUE, echo=FALSE, results = "asis"}
pkgs <- loadedNamespaces()
libTvdB::format_sessioninfo_html(pkgs)
```
