==> analysis/04_01_ER_PESCAn.Rmd <==
---
title: "ER PESCAn: pool our data"
author: 
  - name: "Teun van den Brand"
    email: "t.vd.brand@nki.nl"
    affiliation: "Netherlands Cancer Institute"
date: '`r format(Sys.time(), "%d/%m/%Y")`'
output: 
  html_document:
    theme: united
    highlight: pygments
    fig_caption: yes
    code_folding: hide
    df_print: kable
    toc: true
    toc_depth: 4
    toc_float:
      collapsed: false
bibliography: /DATA/users/t.vd.brand/bibliography/articles.bibtex
knit: |
  # Redirect output to HTML folder
  (function(inputFile, encoding) {
    outputFile <- gsub("Rmd$", "html", basename(inputFile))
    rmarkdown::render(
      input = inputFile,
      encoding = encoding,
      output_file = here::here("html", outputFile)
    )
  })
---

## Setup {.tabset}

<!-- little HTML script to do indentation of the table of contents -->
<script>
    $(document).ready(function() {
      $items = $('div#TOC li');
      $items.each(function(idx) {
        num_ul = $(this).parentsUntil('#TOC').length;
        $(this).css({'text-indent': num_ul * 10, 'padding-left': 0});
      });

    });
</script>

```{css, echo = FALSE}
div.sourceCode {
  overflow-x: hidden;
}
```

### Load packages

```{r stfu, include = FALSE}
# No, I do *not* want your start-up messages, thank you very much
library <- function(...) {
  suppressPackageStartupMessages(base::library(...))
}
```


```{r load_packages}
library(here)
library(ggplot2)
library(scales)
library(GENOVA)
library(data.table)
library(GenomicRanges)
library(rtracklayer)
```

### Load data

```{r load_data}
files <- list.files(
  here("data", "hic", "pools"),
  full.names = TRUE
)
names(files) <- tstrsplit(basename(files), "\\.")[[1]]

peaks <- list.files(
  here("processed_data", "chip", "era"),
  pattern = "_ER.+chr\\.narrowPeak$",
  full.names = TRUE
)

extra <- c(
  "/DATA/users/t.vd.brand/projects/pleural_effusion_2/processed_data/chip/peaks/ER_peaks.narrowPeak"
)
peaks <- c(peaks, extra)


# peaks <- setNames(peaks, gsub("chr\\.narrowPeak$", "", basename(peaks)))
peaks <- setNames(peaks, tstrsplit(basename(peaks), "\\.")[[1]])
# peaks <- peaks[!grepl("liftover.bed$", peaks)]
peaks <- lapply(peaks, import)
# peaks <- setNames(peaks, c("MCF7", "ZR751", "T47D"))
```

### Setup Aesthetics

```{r setup_aes}
mycolour <- "#000000FF" # Opaque Black

theme_set(theme_gray())
theme_update(text = element_text(colour = mycolour),
             line = element_line(colour = mycolour),
             aspect.ratio = 1,
             axis.line  = element_line(colour = mycolour),
             axis.ticks = element_line(colour = mycolour),
             axis.text  = element_text(colour = mycolour),
             legend.key = element_blank(),
             legend.background = element_rect(colour = NA, fill = NA),
             panel.background = element_blank(),
             panel.grid.major = element_line(colour = "grey95"),
             panel.grid.minor = element_blank(),
             plot.background  = element_blank(),
             strip.background = element_blank(),
             strip.text = element_text(colour = mycolour))
CYRUP <- c("#fed500", "#009bef", "#ff5c49", "#949494", "#D8D8D8")
rm(mycolour)
```

## Aim

## Cluster Peaks

```{r}
peaks <- lapply(names(peaks), function(i) {
  x <- peaks[[i]]
  x$celltype <- i
  return(x)
})
suppressWarnings({
  peaks <- do.call(c, peaks)
})
peaks <- peaks[seqnames(peaks) %in% paste0("chr", c(1:22, "X"))]
peaks <- keepSeqlevels(peaks, seqlevelsInUse(peaks))
peaks <- reduce(disjoin(peaks))

peaks$dist_to_next <- c(
  distance(
    peaks[1:(length(peaks) - 1)],
    peaks[2:length(peaks)]
  ), NA
)
dists <- Rle(peaks$dist_to_next < 2e4)
runValue(dists)[is.na(runValue(dists))] <- FALSE
runValue(dists)[runLength(dists) < 3]   <- FALSE
i <- runValue(dists) & !is.na(runValue(dists))
runValue(dists)[i] <- seq_along(runValue(dists)[i])
dists <- pmax(c(decode(dists), 0), c(0, decode(dists)))
peaks$cluster <- dists[-length(dists)]

clusts <- peaks[peaks$cluster != 0 & !is.na(peaks$cluster)]
sclusts <- split(clusts, clusts$cluster)
clusts <- unlist(range(sclusts))
clusts$n_peaks <- lengths(sclusts)
```

## PESCAn

```{r, warning=FALSE}
results <- Map(function(file, nm) {
  
  exp <- load_contacts(file, resolution = 20e3, sample_name = nm)
  
  abs_template <- exp$IDX
  anch <- anchors_PESCAn(
    abs_template, 20000, bed = as.data.frame(clusts),
    dist_thres = c(3e6, Inf)
  )
  
  left  <- as.vector(unname(unclass(anch[, 1])))
  right <- as.vector(unname(unclass(anch[, 2])))
  left  <- abs_template[match(left, V4)]
  right <- abs_template[match(right, V4)]
  rownames(anch) <- paste0(
    left[,  paste0(V1, ":", V2, "-", V3)], ";",
    right[, paste0(V1, ":", V2, "-", V3)]
  )
  
  pescan <- PESCAn(exp, anchors = anch, shift = 1e6, size_bp = 4e5)
  quant  <- quantify(
    pescan, shape = "center_vs_quadrants",
    size = 9, metric = "mean"
  )
  
  # God save the memory
  pescan$signal_raw  <- NULL
  pescan$shifted_raw <- NULL
  
  list(
    pescan = pescan,
    quant  = quant
  )
}, file = files, nm = names(files))

pescans <- lapply(results, `[[`, "pescan")
pescans <- do.call(bundle, pescans)

quants <- lapply(results, `[[`, "quant")
quants <- list(
  per_sample      = rbindlist(lapply(quants, `[[`, "per_sample")),
  per_interaction = rbindlist(lapply(quants, `[[`, "per_interaction"))
)
```

## Results

### Aggregates

```{r}
scans <- reshape2::melt(pescans$obsexp)

ggplot(scans, aes(Var1, Var2, fill = value)) +
  geom_raster() +
  scale_fill_GENOVA_div(
    limits   = range(scans$value[scans$Var3 != "P14_PB"]),
    midpoint = 1, oob = oob_squish,
    name = "Obs/\nExp"
  ) +
  facet_wrap(~ Var3, ncol = 6) +
  scale_x_continuous(
    breaks = c(-250, 0, 250) * 1e3,
    labels = c("-250", 0, "+250"),
    expand = c(0, 0),
    name   = NULL
  ) +
  scale_y_continuous(
    breaks = c(-250, 0, 250) * 1e3,
    labels = c("-250", "0", "+250"),
    expand = c(0, 0),
    name   = NULL
  )

```

### Quantification

```{r}
ggplot(quants$per_sample, aes(foreground - background, sample)) +
  geom_col(width = 2/(1 + sqrt(5))) +
  scale_x_continuous(expand = c(0, 0, 0.1, 0))
```

## Save

```{r}
saveRDS(pescans, here("rds", "01_PESCAn_summaries_all_our_data.rds"))
saveRDS(quants,  here("rds", "01_PESCAn_quantification_all_our_data.rds"))
```


## References

<div id="refs"></div>

## Session Info

```{r session_info, include=TRUE, echo=FALSE, results = "asis"}
pkgs <- loadedNamespaces()
libTvdB::format_sessioninfo_html(pkgs)
```
